refactor - separate dichotomous outputs from structs
