+++ bmds/bmds3/constants.py
@@ -7,6 +7,7 @@ from pydantic import BaseModel
 BMDS_BLANK_VALUE = -9999
 CDF_TABLE_SIZE = 99
 MY_MAX_PARMS = 16
+N_BMD_DIST = 200
 NUM_LIKELIHOODS_OF_INTEREST = 5
 NUM_PRIOR_COLS = 5
 NUM_TESTS_OF_INTEREST = 4
